requirements.txt
alembic==1.9.4
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
python-telegram-bot-pagination==0.0.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
